# Feature/fps jpeg quality optimization