--- a/src/main/java/com/screenai/controller/ScreenSharingApiController.java
+++ b/src/main/java/com/screenai/controller/ScreenSharingApiController.java
@@ -19,23 +19,24 @@ public class ScreenSharingApiController {
 
     @Autowired
     private ScreenCaptureService screenCaptureService;
-    
+
     @Autowired
     private ScreenShareWebSocketHandler webSocketHandler;
 
     /**
      * Get current screen sharing status and statistics
+     * 
      * @return Status information including capture method, viewer count, etc.
      */
     @GetMapping("/api/status")
     public ResponseEntity<Map<String, Object>> getStatus() {
         Map<String, Object> status = new HashMap<>();
-        
+
         // Basic status
         status.put("initialized", screenCaptureService.isInitialized());
         status.put("capturing", screenCaptureService.isCapturing());
         status.put("captureMethod", screenCaptureService.getCaptureMethod());
-        
+
         // Screen information
         if (screenCaptureService.getScreenBounds() != null) {
             Map<String, Object> screen = new HashMap<>();
@@ -43,30 +44,39 @@ public class ScreenSharingApiController {
             screen.put("height", screenCaptureService.getScreenBounds().height);
             status.put("screenResolution", screen);
         }
-        
+
         // Viewer information
         status.put("viewerCount", webSocketHandler.getViewerCount());
-        
+
+        // Performance information
+        if (screenCaptureService.isInitialized()) {
+            Map<String, Object> performance = new HashMap<>();
+            performance.put("fps", screenCaptureService.getCurrentFPS());
+            performance.put("jpegQuality", screenCaptureService.getCurrentJPEGQuality());
+            performance.put("optimizations", screenCaptureService.getOptimizationStatus());
+            status.put("performance", performance);
+        }
+
         // System information
         status.put("serverTime", System.currentTimeMillis());
         status.put("osName", System.getProperty("os.name"));
         status.put("javaVersion", System.getProperty("java.version"));
-        
+
         return ResponseEntity.ok(status);
     }
-    
+
     /**
      * Start screen capture manually
      */
     @GetMapping("/api/start-capture")
     public ResponseEntity<Map<String, Object>> startCapture() {
         Map<String, Object> response = new HashMap<>();
-        
+
         try {
             if (!screenCaptureService.isInitialized()) {
                 screenCaptureService.initialize();
             }
-            
+
             if (screenCaptureService.isInitialized() && !screenCaptureService.isCapturing()) {
                 screenCaptureService.startCapture();
                 response.put("success", true);
@@ -84,17 +94,17 @@ public class ScreenSharingApiController {
             response.put("success", false);
             response.put("message", "Error starting capture: " + e.getMessage());
         }
-        
+
         return ResponseEntity.ok(response);
     }
-    
+
     /**
      * Stop screen capture manually
      */
     @GetMapping("/api/stop-capture")
     public ResponseEntity<Map<String, Object>> stopCapture() {
         Map<String, Object> response = new HashMap<>();
-        
+
         try {
             screenCaptureService.stopCapture();
             response.put("success", true);
@@ -103,7 +113,7 @@ public class ScreenSharingApiController {
             response.put("success", false);
             response.put("message", "Error stopping capture: " + e.getMessage());
         }
-        
+
         return ResponseEntity.ok(response);
     }
 }