--- a/src/main/java/com/screenai/controller/WebController.java
+++ b/src/main/java/com/screenai/controller/WebController.java
@@ -1,4 +1,4 @@
-package com.aiscreensharing.controller;
+package com.screenai.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,18 +8,20 @@ import org.springframework.web.bind.annotation.GetMapping;
  */
 @Controller
 public class WebController {
-    
+
     /**
      * Serves the main screen sharing viewer page
+     * 
      * @return the name of the Thymeleaf template
      */
     @GetMapping("/")
     public String index() {
         return "index";
     }
-    
+
     /**
      * Alternative endpoint for the viewer
+     * 
      * @return the name of the Thymeleaf template
      */
     @GetMapping("/viewer")