--- a/src/main/java/com/screenai/service/ScreenCaptureService.java
+++ b/src/main/java/com/screenai/service/ScreenCaptureService.java
@@ -19,12 +19,19 @@ import org.bytedeco.javacv.Java2DFrameConverter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import com.screenai.handler.ScreenShareWebSocketHandler;
 
+import javax.imageio.IIOImage;
+import javax.imageio.ImageWriteParam;
+import javax.imageio.ImageWriter;
+import javax.imageio.stream.ImageOutputStream;
+import java.util.Iterator;
+
 /**
- * Screen Capture Service using JavaCV  for real screen capture
+ * Screen Capture Service using JavaCV for real screen capture
  * Implements FFmpegFrameGrabber for platform-specific screen capture
  * No AWT Robot fallback - JavaCV only implementation
  */
@@ -32,7 +39,18 @@ import com.screenai.handler.ScreenShareWebSocketHandler;
 public class ScreenCaptureService {
 
     private static final Logger logger = LoggerFactory.getLogger(ScreenCaptureService.class);
-    private static final int FRAME_RATE = 10; // 10 FPS for balance between performance and smoothness
+
+    @Value("${screenai.capture.fps:15}")
+    private int frameRate;
+
+    @Value("${screenai.capture.jpeg.quality:70}")
+    private int jpegQuality;
+
+    @Value("${screenai.capture.optimize.ultrafast:true}")
+    private boolean ultraFastMode;
+
+    @Value("${screenai.capture.optimize.zerolatency:true}")
+    private boolean zeroLatencyMode;
 
     @Autowired
     private ScreenShareWebSocketHandler webSocketHandler;
@@ -53,15 +71,15 @@ public class ScreenCaptureService {
             GraphicsEnvironment ge = GraphicsEnvironment.getLocalGraphicsEnvironment();
             GraphicsDevice gd = ge.getDefaultScreenDevice();
             screenRect = gd.getDefaultConfiguration().getBounds();
-            
+
             logger.info("Detected screen resolution: {}x{}", screenRect.width, screenRect.height);
-            
+
             // Initialize JavaCV Frame Converter
             converter = new Java2DFrameConverter();
-            
+
             // Initialize JavaCV screen capture
             initializePlatformScreenCapture();
-            
+
             if (frameGrabber != null) {
                 try {
                     frameGrabber.start();
@@ -74,20 +92,20 @@ public class ScreenCaptureService {
             } else {
                 throw new RuntimeException("JavaCV screen capture not available for this platform");
             }
-            
+
         } catch (Exception e) {
             logger.error("Failed to initialize JavaCV screen capture", e);
             throw new RuntimeException("Screen capture initialization failed", e);
         }
     }
-    
+
     /**
      * Initializes platform-specific screen capture using FFmpeg
      */
     private void initializePlatformScreenCapture() {
         String osName = System.getProperty("os.name").toLowerCase();
         logger.info("Initializing JavaCV screen capture for OS: {}", osName);
-        
+
         try {
             if (osName.contains("windows")) {
                 // Windows: Use gdigrab
@@ -95,16 +113,16 @@ public class ScreenCaptureService {
                 frameGrabber.setFormat("gdigrab");
                 frameGrabber.setImageWidth(screenRect.width);
                 frameGrabber.setImageHeight(screenRect.height);
-                frameGrabber.setFrameRate(FRAME_RATE);
+                frameGrabber.setFrameRate(frameRate);
                 logger.info("Windows gdigrab screen capture configured");
-                
+
             } else if (osName.contains("mac")) {
                 // macOS: Try multiple approaches
                 try {
                     // First attempt: Use desktop capture
                     frameGrabber = new FFmpegFrameGrabber("Capture screen 0");
                     frameGrabber.setFormat("avfoundation");
-                    frameGrabber.setFrameRate(FRAME_RATE);
+                    frameGrabber.setFrameRate(frameRate);
                     logger.info("macOS avfoundation desktop capture configured");
                 } catch (Exception e1) {
                     logger.debug("Primary macOS capture failed: {}", e1.getMessage());
@@ -112,48 +130,53 @@ public class ScreenCaptureService {
                         // Second attempt: Use desktop index
                         frameGrabber = new FFmpegFrameGrabber("1:none");
                         frameGrabber.setFormat("avfoundation");
-                        frameGrabber.setFrameRate(FRAME_RATE);
+                        frameGrabber.setFrameRate(frameRate);
                         logger.info("macOS avfoundation index capture configured");
                     } catch (Exception e2) {
                         logger.debug("Secondary macOS capture failed: {}", e2.getMessage());
                         frameGrabber = null;
                     }
                 }
-                
+
             } else if (osName.contains("linux")) {
                 // Linux: Use x11grab
                 String display = System.getenv("DISPLAY");
-                if (display == null) display = ":0.0";
-                
+                if (display == null)
+                    display = ":0.0";
+
                 frameGrabber = new FFmpegFrameGrabber(display);
                 frameGrabber.setFormat("x11grab");
                 frameGrabber.setImageWidth(screenRect.width);
                 frameGrabber.setImageHeight(screenRect.height);
-                frameGrabber.setFrameRate(FRAME_RATE);
+                frameGrabber.setFrameRate(frameRate);
                 logger.info("Linux x11grab screen capture configured for display: {}", display);
-                
+
             } else {
                 logger.warn("Unsupported operating system for JavaCV screen capture: {}", osName);
                 frameGrabber = null;
                 return;
             }
-            
+
             // Set common options if frameGrabber is available
             if (frameGrabber != null) {
                 try {
-                    frameGrabber.setOption("preset", "ultrafast");
-                    frameGrabber.setOption("tune", "zerolatency");
+                    if (ultraFastMode) {
+                        frameGrabber.setOption("preset", "ultrafast");
+                    }
+                    if (zeroLatencyMode) {
+                        frameGrabber.setOption("tune", "zerolatency");
+                    }
                 } catch (Exception optionError) {
                     logger.debug("Could not set FFmpeg options: {}", optionError.getMessage());
                 }
             }
-            
+
         } catch (Exception e) {
             logger.warn("Failed to configure JavaCV screen capture: {}", e.getMessage());
             frameGrabber = null;
         }
     }
-    
+
     /**
      * Starts continuous screen capture and broadcasting
      */
@@ -162,10 +185,10 @@ public class ScreenCaptureService {
             logger.warn("Cannot start capture - not initialized or already capturing");
             return;
         }
-        
+
         isCapturing = true;
         scheduler = Executors.newScheduledThreadPool(1);
-        
+
         scheduler.scheduleAtFixedRate(() -> {
             try {
                 String frameData = captureScreenAsBase64();
@@ -178,11 +201,11 @@ public class ScreenCaptureService {
             } catch (Exception e) {
                 logger.error("Error during screen capture broadcast", e);
             }
-        }, 0, 1000 / FRAME_RATE, TimeUnit.MILLISECONDS);
-        
-        logger.info("Screen capture started at {} FPS (using JavaCV)", FRAME_RATE);
+        }, 0, 1000 / frameRate, TimeUnit.MILLISECONDS);
+
+        logger.info("Screen capture started at {} FPS (using JavaCV)", frameRate);
     }
-    
+
     /**
      * Stops continuous screen capture
      */
@@ -193,9 +216,11 @@ public class ScreenCaptureService {
             logger.info("Screen capture stopped");
         }
     }
-    
+
     /**
      * Captures the current screen and returns it as a Base64 encoded JPEG string
+     * with configurable quality
+     * 
      * @return Base64 encoded JPEG image of the screen
      */
     public String captureScreenAsBase64() {
@@ -204,22 +229,58 @@ public class ScreenCaptureService {
             if (screenCapture == null) {
                 return null;
             }
-            
-            // Convert to JPEG and encode as Base64
+
+            // Convert to JPEG with configurable quality and encode as Base64
             ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            ImageIO.write(screenCapture, "jpg", baos);
+            writeJPEGWithQuality(screenCapture, baos, jpegQuality);
             byte[] imageBytes = baos.toByteArray();
-            
+
             return Base64.getEncoder().encodeToString(imageBytes);
-            
+
         } catch (IOException e) {
             logger.error("Failed to capture screen", e);
             return null;
         }
     }
-    
+
+    /**
+     * Writes BufferedImage as JPEG with specified quality
+     * 
+     * @param image   The BufferedImage to write
+     * @param output  The output stream to write to
+     * @param quality Quality level (0-100, where 100 is highest quality)
+     */
+    private void writeJPEGWithQuality(BufferedImage image, ByteArrayOutputStream output, int quality)
+            throws IOException {
+        // Validate quality range
+        quality = Math.max(30, Math.min(100, quality)); // Clamp between 30-100
+
+        Iterator<ImageWriter> writers = ImageIO.getImageWritersByFormatName("jpg");
+        if (!writers.hasNext()) {
+            throw new IOException("No JPEG writer available");
+        }
+
+        ImageWriter writer = writers.next();
+        ImageWriteParam param = writer.getDefaultWriteParam();
+
+        if (param.canWriteCompressed()) {
+            param.setCompressionMode(ImageWriteParam.MODE_EXPLICIT);
+            param.setCompressionQuality(quality / 100.0f); // Convert to 0.0-1.0 range
+        }
+
+        try (ImageOutputStream ios = ImageIO.createImageOutputStream(output)) {
+            writer.setOutput(ios);
+            writer.write(null, new IIOImage(image, null, null), param);
+        } finally {
+            writer.dispose();
+        }
+
+        logger.debug("JPEG encoded with quality: {}%", quality);
+    }
+
     /**
      * Captures the current screen as a BufferedImage
+     * 
      * @return BufferedImage of the current screen
      */
     public BufferedImage captureScreen() {
@@ -227,10 +288,10 @@ public class ScreenCaptureService {
             logger.debug("Screen capture not initialized");
             return null;
         }
-        
+
         return captureScreenWithJavaCV();
     }
-    
+
     /**
      * Captures screen using JavaCV FFmpegFrameGrabber
      */
@@ -239,7 +300,7 @@ public class ScreenCaptureService {
             logger.debug("JavaCV frame grabber not available");
             return null;
         }
-        
+
         try {
             // Capture frame using JavaCV FFmpegFrameGrabber
             Frame frame = frameGrabber.grab();
@@ -252,25 +313,26 @@ public class ScreenCaptureService {
                 logger.debug("No frame captured from JavaCV");
                 return null;
             }
-            
+
         } catch (Exception e) {
             logger.error("JavaCV screen capture failed: {}", e.getMessage());
             return null;
         }
     }
-    
+
     /**
      * Captures screen as JavaCV Frame for advanced processing
+     * 
      * @return org.bytedeco.javacv.Frame object
      */
     public Frame captureScreenAsFrame() {
         if (!isInitialized) {
             return null;
         }
-        
+
         return captureFrameWithJavaCV();
     }
-    
+
     /**
      * Captures frame directly with JavaCV
      */
@@ -278,7 +340,7 @@ public class ScreenCaptureService {
         if (frameGrabber == null) {
             return null;
         }
-        
+
         try {
             Frame frame = frameGrabber.grab();
             return frame;
@@ -287,41 +349,46 @@ public class ScreenCaptureService {
             return null;
         }
     }
-    
+
     /**
      * Gets the screen dimensions
+     * 
      * @return Rectangle representing screen bounds
      */
     public Rectangle getScreenBounds() {
         return screenRect;
     }
-    
+
     /**
      * Gets the JavaCV frame converter
+     * 
      * @return Java2DFrameConverter instance
      */
     public Java2DFrameConverter getConverter() {
         return converter;
     }
-    
+
     /**
      * Checks if screen capture is properly initialized
+     * 
      * @return true if screen capture is working
      */
     public boolean isInitialized() {
         return isInitialized;
     }
-    
+
     /**
      * Checks if screen capture is currently running
+     * 
      * @return true if actively capturing and broadcasting
      */
     public boolean isCapturing() {
         return isCapturing;
     }
-    
+
     /**
      * Gets the current capture method being used
+     * 
      * @return String describing the capture method
      */
     public String getCaptureMethod() {
@@ -330,13 +397,41 @@ public class ScreenCaptureService {
         }
         return "JavaCV FFmpeg";
     }
-    
+
+    /**
+     * Gets the current FPS setting
+     * 
+     * @return Current frames per second
+     */
+    public int getCurrentFPS() {
+        return frameRate;
+    }
+
+    /**
+     * Gets the current JPEG quality setting
+     * 
+     * @return Current JPEG quality (30-100)
+     */
+    public int getCurrentJPEGQuality() {
+        return jpegQuality;
+    }
+
+    /**
+     * Gets performance optimization status
+     * 
+     * @return String describing current optimizations
+     */
+    public String getOptimizationStatus() {
+        return String.format("UltraFast: %s, ZeroLatency: %s, FPS: %d, Quality: %d%%",
+                ultraFastMode, zeroLatencyMode, frameRate, jpegQuality);
+    }
+
     /**
      * Cleanup resources when service is destroyed
      */
     public void cleanup() {
         stopCapture();
-        
+
         if (frameGrabber != null) {
             try {
                 frameGrabber.stop();
@@ -346,7 +441,7 @@ public class ScreenCaptureService {
                 logger.warn("Error during JavaCV cleanup: {}", e.getMessage());
             }
         }
-        
+
         logger.info("Screen capture service cleaned up");
     }
 }