--- a/src/main/resources/application.properties
+++ b/src/main/resources/application.properties
@@ -29,3 +29,18 @@ info.app.version=1.0.0
 # Maximum message size (for large screen captures)
 spring.websocket.max-text-message-buffer-size=1048576
 spring.websocket.max-binary-message-buffer-size=1048576
+
+# Screen Capture Performance Configuration
+# FPS Settings (5-30 range for performance optimization)
+screenai.capture.fps=15
+screenai.capture.fps.min=5
+screenai.capture.fps.max=30
+
+# JPEG Quality Settings (30-100 range, higher = better quality but larger files)
+screenai.capture.jpeg.quality=70
+screenai.capture.jpeg.quality.min=30
+screenai.capture.jpeg.quality.max=100
+
+# Performance optimization options
+screenai.capture.optimize.ultrafast=true
+screenai.capture.optimize.zerolatency=true