## 🚀 Performance Optimization: FPS & JPEG Quality Configuration

### 📈 **Improvements**
- **FPS Increase**: 10 FPS → 15 FPS (50% smoother streaming)
- **JPEG Quality Control**: Configurable compression (70% default for optimal balance)
- **Dynamic Configuration**: All settings configurable via `application.properties`
- **FFmpeg Optimization**: UltraFast preset + Zero latency tuning

### 🔧 **Technical Changes**
- ✅ Added `@Value` Spring configuration injection for dynamic settings
- ✅ Implemented custom JPEG quality control with `ImageWriter` API
- ✅ Enhanced `/api/status` endpoint with performance metrics
- ✅ Added runtime performance monitoring methods
- ✅ Configurable FFmpeg optimization flags

### 📊 **Performance Impact**
- **Reduced Lag**: 33% reduction in frame update delay (15 vs 10 FPS)
- **Bandwidth Efficiency**: Optimized JPEG compression
- **Production Ready**: Settings tunable without code changes
- **Monitoring**: Real-time performance metrics via API

### 🧪 **Testing**
- ✅ Application starts successfully with new configuration
- ✅ FPS increase confirmed: `Stream #0:0: Video: bmp, bgra, 1536x864, 637016 kb/s, 15 fps`
- ✅ Performance metrics available via `/api/status`
- ✅ All existing functionality preserved

### 📁 **Files Modified**
- `src/main/java/com/screenai/service/ScreenCaptureService.java`
- `src/main/resources/application.properties`
- `src/main/java/com/screenai/controller/ScreenSharingApiController.java`

### 🎯 **Jira Task**
Addresses: Configuration of FPS & JPEG quality to reduce screen lag